# semester1/week4/ginserver1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "这是一个基于Gin框架的用户管理系统API，提供用户注册、登录、个人信息管理等功能",
    "title": "用户管理API",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded"
  ],
  "produces": [
    "application/json",
    "text/html"
  ],
  "paths": {
...
    "/login": {
      "get": {
        "description": "返回用户登录页面",
        "produces": [
          "text/html"
        ],
        "tags": [
          "页面"
        ],
        "operationId": "loginPage",
        "responses": {
          "200": {
            "description": "成功返回登录页面"
          }
        }
      },
      "post": {
        "description": "用户登录并返回JWT令牌",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户"
        ],
        "operationId": "loginUser",
        "parameters": [
          {
            "type": "string",
            "description": "用户名",
            "name": "username",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "密码",
            "name": "password",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "302": {
            "description": "登录成功，重定向到个人资料页面"
          },
          "400": {
            "description": "用户不存在或信息为空",
            "schema": {
              "$ref": "#/definitions/APIResponse"
            }
          },
          "403": {
            "description": "密码错误",
            "schema": {
              "$ref": "#/definitions/APIResponse"
            }
          },
          "500": {
            "description": "服务器内部错误",
            "schema": {
              "$ref": "#/definitions/APIResponse"
            }
          }
        }
      }
    },
...
    "/register": {
      "get": {
        "description": "返回用户注册页面",
        "produces": [
          "text/html"
        ],
        "tags": [
          "页面"
        ],
        "operationId": "registerPage",
        "responses": {
          "200": {
            "description": "成功返回注册页面"
          }
        }
      },
      "post": {
        "description": "注册新用户",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户"
        ],
        "operationId": "registerUser",
        "parameters": [
          {
            "type": "string",
            "description": "用户名",
            "name": "username",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "密码",
            "name": "password",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "姓名",
            "name": "name",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "注册成功",
            "schema": {
              "$ref": "#/definitions/APIResponse"
            }
          },
          "400": {
            "description": "用户已存在或信息为空",
            "schema": {
              "$ref": "#/definitions/APIResponse"
            }
          },
          "500": {
            "description": "服务器内部错误",
            "schema": {
              "$ref": "#/definitions/APIResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "APIResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        },
        "data": {
          "type": "object"
        }
      }
    }
  },
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT认证令牌，格式为\"Bearer {token}\""
    }
  }
}